requirements.txt
asgiref==3.9.1
defusedxml==0.7.1
diff-match-patch==20241021
Django==4.2.1
django-import-export==3.2.0
et_xmlfile==2.0.0
odfpy==1.4.1
openpyxl==3.1.2
PyYAML==6.0.2
sqlparse==0.5.3
tablib==3.8.0
xlrd==2.0.2
xlwt==1.3.0
